add/server: finish controller lesson with image and video

+++ server/src/main/java/com/duokoala/server/dto/response/mediaResponse/VideoResponse.java
@@ -11,5 +11,5 @@ import lombok.experimental.FieldDefaults;
 public class VideoResponse {
     String videoId;
     String videoUrl;
-    int videoDuration;
+    double videoDuration;
 }
